Add files via upload
This module counts the number of consecutive DM_INPUT value parameters that are NOT zero and uses that count as the number of list items to display on a Smart Graphics dynamic button list.  DM source input numbers are mapped to the source list.  DM Video sync is used to Enable/Disable source list buttons.